:heavy_plus_sign: :wrench: [feat]: add react navigation configuration and dependencies
